Solucionado el bug en el cálculo de las distancias con las coordenadas

+++ kolombo/src/configuration.h
@@ -28,7 +28,7 @@ class configuration
 		QString nombreDB, hostDB, passDB, driverDB, userDB;
 		QString nombreP, dniP, clubP, poblacionP, licenciaP, provinciaP, calleP, numeroP;
 		bool connectOnStart;
-        double coordX, coordY;
+        long int coordX, coordY;
 		int ultimaAnilla;
 		int anillaSeleccion;
